pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.8.4",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "streamlit>=1.42.0",
]
